+++ gpu-video/examples/player/player/decoder.rs
@@ -9,12 +9,13 @@ pub fn run_decoder(
     tx: SyncSender<super::FrameWithPts>,
     framerate: u64,
     device: wgpu::Device,
+    queue: wgpu::Queue,
     mut bytestream_reader: impl Read,
 ) {
     let mut decoder = device
         .video()
         .unwrap()
-        .create_wgpu_textures_decoder_h264(DecoderParameters::default())
+        .create_wgpu_textures_decoder_h264(&queue, DecoderParameters::default())
         .unwrap();
     let frame_interval = 1.0 / (framerate as f64);
     let mut frame_number = 0u64;
@@ -33,7 +34,7 @@ pub fn run_decoder(
 
     while let Ok(n) = bytestream_reader.read(&mut buffer) {
         if n == 0 {
-            return;
+            break;
         }
 
         let frame = EncodedInputChunk {