+++ gpu-video/src/exports.rs
@@ -1,8 +1,8 @@
 pub mod parameters {
     pub use crate::adapter::VideoAdapterDescriptor;
     pub use crate::device::{
-        ColorRange, ColorSpace, DecoderParameters, EncoderOutputParameters, EncoderParametersH264,
-        EncoderParametersH265, MissedFrameHandling, Rational, VideoDeviceDescriptor,
+        ColorRange, ColorSpace, CorruptedStateHandling, DecoderParameters, EncoderOutputParameters,
+        EncoderParametersH264, EncoderParametersH265, Rational, VideoDeviceDescriptor,
         VideoParameters,
     };
     pub use crate::instance::VideoInstanceDescriptor;
@@ -118,7 +118,7 @@ use crate::device::{
 };
 use crate::parameters::{H264Profile, H265Profile, RateControl};
 use crate::parser::h264::AccessUnit;
-use std::sync::Arc;
+use std::sync::{Arc, Mutex};
 
 #[cfg(feature = "wgpu")]
 pub use crate::{
@@ -130,8 +130,8 @@ pub use crate::{
 
 pub use crate::adapter::VideoAdapter;
 #[cfg(feature = "wgpu")]
-pub use crate::decoders::WgpuTexturesDecoder;
-pub use crate::decoders::{BytesDecoder, VideoDecoderError};
+pub use crate::decoders::WgpuTexturesDecoderH264;
+pub use crate::decoders::{BytesDecoderH264, VideoDecoderError};
 pub use crate::encoders::{BytesEncoderH264, BytesEncoderH265, VideoEncoderError};
 #[cfg(feature = "wgpu")]
 pub use crate::encoders::{WgpuTexturesEncoderH264, WgpuTexturesEncoderH265};
@@ -182,25 +182,34 @@ impl std::fmt::Debug for VideoDevice {
 }
 
 impl VideoDevice {
-    pub fn create_bytes_decoder_h264(
+    pub fn create_bytes_decoder_h264<F>(
         &self,
         parameters: DecoderParameters,
-    ) -> Result<BytesDecoder, VideoDecoderError> {
-        self.inner.clone().create_bytes_decoder_h264(parameters)
+        on_frame: F,
+    ) -> Result<BytesDecoderH264, VideoDecoderError>
+    where
+        F: FnMut(OutputFrame<RawFrameData>) + Send + 'static,
+    {
+        self.inner
+            .clone()
+            .create_bytes_decoder_h264(parameters, Arc::new(Mutex::new(on_frame)))
     }
 
     #[cfg(feature = "wgpu")]
     pub fn create_wgpu_textures_decoder_h264(
         &self,
+        wgpu_queue: &wgpu::Queue,
         parameters: DecoderParameters,
-    ) -> Result<WgpuTexturesDecoder, VideoDecoderError> {
+    ) -> Result<WgpuTexturesDecoderH264, VideoDecoderError> {
         let Some(wgpu_device) = self.wgpu_device.clone() else {
             return Err(VideoDecoderError::VideoDeviceWithoutWgpu);
         };
 
-        self.inner
-            .clone()
-            .create_wgpu_textures_decoder_h264(wgpu_device, parameters)
+        self.inner.clone().create_wgpu_textures_decoder_h264(
+            wgpu_device,
+            wgpu_queue.clone(),
+            parameters,
+        )
     }
 
     /// Create a single-input multiple-output transcoder.
@@ -401,7 +410,7 @@ pub enum DecoderEvent<'a, ParsedFrame> {
 
     /// Signal the decoder that a chunk of the bitstream was lost.
     ///
-    /// What the decoder will do depends on the set [`parameters::MissedFrameHandling`]
+    /// What the decoder will do depends on the set [`parameters::CorruptedStateHandling`]
     SignalDataLoss,
 
     /// Flush all frames from the decoder.
@@ -428,6 +437,7 @@ pub struct InputFrame<T> {
 }
 
 /// Additional information about the decoded frame.
+#[derive(Clone)]
 pub struct FrameMetadata {
     pub pts: Option<u64>,
     pub color_space: ColorSpace,